+++ scripts/release.sh
@@ -1,9 +1,9 @@
 #!/bin/bash -e
 
-cd $(dirname $0)
+cd $(dirname $0)/..
 
 # Perform a build
-./build.sh
+./scripts/build.sh
 
 # The tag is something like: v1.2.3
 export TAG="$(head -n1 CHANGELOG.md | grep -E -o 'v[^ ]+')"
@@ -48,3 +48,28 @@ github-release upload \
     --name "notable-${TAG}.linux-amd64.zip" \
     --file target/notable-${TAG}.linux-amd64.zip
 
+# Docker vars
+project=jmcfarlane
+binary=notable
+build_tag=$project/${binary}-build
+run_tag=$project/$binary
+
+# Build the binary via docker
+docker build --no-cache -t $build_tag -f Dockerfile.build .
+
+# Copy out the (musl) binary
+docker run --rm -v $(pwd):/mount $build_tag cp \
+    /go/src/github.com/$project/$binary/target/$binary-${TAG}.linux-amd64/$binary \
+    /mount/$binary
+
+# Build the runnable container
+docker build --no-cache -t $run_tag .
+docker tag $run_tag:latest $run_tag:$TAG
+
+# Include some info about the containers
+docker images $project/$binary*
+
+# Pubish
+docker login
+docker push jmcfarlane/notable:latest
+docker push jmcfarlane/notable:$TAG